split taxonomy and taxonomy search
